+++ Mozu.Api/Contracts/Content/PropertyTypeCollection.cs
@@ -19,24 +19,24 @@ namespace Mozu.Api.Contracts.Content
 		public class PropertyTypeCollection
 		{
 			///
-			///The number of pages returned based on the startIndex and pageSize values specified. This value is system-supplied and read-only.
+			///The number of pages returned based on the startIndex and pageSize supplied. Signed 64-bit (8-byte) integer. System-supplied and read-only.
 			///
-			public int PageCount { get; set; }
+			public long PageCount { get; set; }
 
 			///
-			///Specifies the number of results to display on each page when creating paged results from a query. The maximum value is 200.
+			///Used to create paged results from a query. Specifies the number of results to display on each page. Maximum: 200.
 			///
 			public int PageSize { get; set; }
 
 			///
-			///Indicates the zero-based offset in the complete result set where the returned entities begin, when creating paged results from a query. For example, with a PageSize of 25, to get the 51st through the 75th items, use startIndex=3.
+			///"Used to create paged results from a query. Indicates the zero-based offset in the complete result set where the returned entities begin. For example, with a PageSize of 25, to get the 51st through the 75th items, use startIndex=3."
 			///
 			public int StartIndex { get; set; }
 
 			///
-			///The number of results listed in the query collection, represented by a signed 64-bit (8-byte) integer. This value is system-supplied and read-only.
+			///The number of the resource entity items listed in the query collection. Signed 64-bit (8-byte) integer. System-supplied and read-only.
 			///
-			public int TotalCount { get; set; }
+			public long TotalCount { get; set; }
 
 			///
 			///An array list of objects in the returned collection.
