+++ Mozu.Api/Contracts/ProductAdmin/AttributeCollection.cs
@@ -14,26 +14,26 @@ using System.Collections.Generic;
 namespace Mozu.Api.Contracts.ProductAdmin
 {
 		///
-		///	Paged list of defined product attributes.
+		///	Paged list of properties that are assigned to a product.
 		///
 		public class AttributeCollection
 		{
 			///
-			///The number of pages returned based on the startIndex and pageSize values specified. This value is system-supplied and read-only.
+			///The number of pages returned based on the startIndex and pageSize supplied. Signed 64-bit (8-byte) integer. System-supplied and read-only.
 			///
-			public int PageCount { get; set; }
+			public long PageCount { get; set; }
 
 			///
-			///Specifies the number of results to display on each page when creating paged results from a query. The maximum value is 200.
+			///Used to create paged results from a query. Specifies the number of results to display on each page. Maximum: 200.
 			///
 			public int PageSize { get; set; }
 
 			public int StartIndex { get; set; }
 
 			///
-			///The number of results listed in the query collection, represented by a signed 64-bit (8-byte) integer. This value is system-supplied and read-only.
+			///The number of the resource entity items listed in the query collection. Signed 64-bit (8-byte) integer. System-supplied and read-only.
 			///
-			public int TotalCount { get; set; }
+			public long TotalCount { get; set; }
 
 			///
 			///An array list of objects in the returned collection.
